index/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==3.0
djangorestframework==3.10.1
kombu==4.6.3
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
python-crontab==2.3.8
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2019.1
redis==3.2.1
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
